# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SQL Practice API
  version: "1.0.0"
  description: |
    API for the SQL Practice app (interactive SQL tasks checked against a real
    PostgreSQL database, sessions/learning plans, login/role system).

    **This spec is hand-maintained, not generated from code.** It documents a
    first, deliberately partial slice of the API (see "Phase 1" scope below) —
    if anything here looks stale compared to the actual route code, trust the
    code and update this file.

    ### Authentication
    This API uses a session cookie (`connect.sid`, set by `express-session` +
    `connect-pg-simple`), **not** bearer tokens. To use "Try it out" below:
    1. Call `POST /api/auth/login` with a real username/password — a valid
       session cookie is stored by your browser automatically.
    2. Every subsequent "Try it out" call from this same page reuses that
       cookie automatically — no token to copy/paste anywhere.
    3. Call `POST /api/auth/logout` to end the session.

    This only works because Swagger UI is served by the backend itself
    (same origin as `/api/*`) — see `src/routes/docs.js`.

    ### Phase 1 scope
    Only a first, "safe to explore" slice of the API is documented here:
    health, auth, datasets, tasks (read + solution), the query sandbox, and
    the table browser. Endpoints with heavier role/ownership matrices
    (sessions, users, progress, mentor assignments, mentor students) are
...
    full API surface in the meantime.
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: >-
        Session cookie set by POST /api/auth/login. httpOnly, sameSite=lax,
        secure in production. There is no bearer-token alternative.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
      required: [error]
      example:
        error: "Not authenticated."
...
    TableColumn:
      type: object
      properties:
        column_name:
          type: string
        data_type:
          type: string
        is_nullable:
          type: string
          enum: [YES, NO]
        column_default:
          type: string
          nullable: true
      example:
        column_name: id
        data_type: integer
        is_nullable: "NO"
        column_default: null
    TablePreview:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
        columns:
          type: array
          items:
            type: string
        rowCount:
          type: integer
          description: Total row count in the table (not just the previewed page).
      required: [rows, columns, rowCount]
...
security:
  - sessionCookie: []
paths:
...
  /api/tables:
    get:
      tags: [Tables]
      summary: List tables in the active practice schema
      description: >-
        Requires login. If `sessionId` is given, the schema is resolved from
        that session and re-authorized via `canAccessStudent` (admin: any;
        mentor: their own or an assigned student's; student: only their own).
        Without `sessionId`, falls back to the default academic schema.
      operationId: listTables
      x-visibility: authenticated
      parameters:
        - name: sessionId
          in: query
          schema: { type: integer }
          description: Optional. Determines which dataset's schema to browse.
      responses:
        "200":
          description: Table names in the resolved schema.
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
              example: ["students", "exams", "professors"]
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Not authorized to view this session's tables.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: sessionId does not refer to an existing session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/tables/{tableName}/columns:
    get:
      tags: [Tables]
      summary: Get column metadata for a table
      description: Same auth/schema-resolution rules as GET /api/tables.
      operationId: getTableColumns
      x-visibility: authenticated
      parameters:
        - name: tableName
          in: path
          required: true
          schema: { type: string, pattern: "^[a-zA-Z_][a-zA-Z0-9_]*$" }
          example: students
        - name: sessionId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Column metadata, in ordinal position order.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TableColumn" }
        "400":
          description: Invalid table name.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Not authorized to view this session's tables, or the table doesn't exist in the resolved schema.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: sessionId does not refer to an existing session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/tables/{tableName}/preview:
    get:
      tags: [Tables]
      summary: Preview a table's data (first 50 rows + total count)
      description: Same auth/schema-resolution rules as GET /api/tables.
      operationId: getTablePreview
      x-visibility: authenticated
      parameters:
        - name: tableName
          in: path
          required: true
          schema: { type: string, pattern: "^[a-zA-Z_][a-zA-Z0-9_]*$" }
          example: students
        - name: sessionId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Up to 50 rows, plus the table's total row count.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TablePreview" }
        "400":
          description: Invalid table name.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Not authorized to view this session's tables, or the table doesn't exist in the resolved schema.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: sessionId does not refer to an existing session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...